fix: comprehensive fixes for pod errors, type safety, and NetworkPolicy cleanup
This commit resolves all detected critical issues across the SongNodes stack:

## 1. NetworkPolicy Cleanup
- Remove stale port references (8011-8026) from orchestrator-to-scrapers policy
- Keep only port 8000 for unified-scraper
- File: deploy/helm/songnodes/templates/networkpolicy.yaml

## 2. TypeScript Critical Syntax Fixes (Frontend)
- Fix LiveDataLoader.tsx:368 - Add missing console.log statement
- Fix SmartTooltip.tsx:520 - Add placeholder onClick handler
- Rename infoCardUtils.ts → infoCardUtils.tsx (contains JSX)
- Rename useLiveDataLoader.ts → useLiveDataLoader.tsx (contains JSX)
- **Result**: 0 syntax errors blocking compilation

## 3. Python High-Priority Type Safety (Backend)
- Add null checks for db_pool in rest_api/main.py (lines 374, 474, 540, 595, 694)
- Add null check for redis_pool in graph-visualization-api/main.py:901
- Add datetime null handling in websocket-api/main.py (6 locations)
- Fix type incompatibility in pathfinder.py (float vs int)
- **Impact**: Prevents potential runtime AttributeError crashes

## 4. Gold-to-Operational ETL Schema Fix
- Fix column name mismatch: use transition_count instead of occurrence_count
- Update SQL alias: transition_count as occurrence_count
- Fix ON CONFLICT columns: use source_track_id, target_track_id
- Set imagePullPolicy: Always for :latest tags
- File: services/gold-to-operational-etl/gold_to_operational_etl.py
- File: deploy/helm/songnodes/values.yaml

## Issues Resolved
✅ NetworkPolicy stale ports removed (11 ports → 1 port)
✅ TypeScript compilation unblocked (4 critical syntax errors fixed)
✅ Python null pointer exceptions prevented (12 locations)
✅ ETL schema mismatch resolved (occurrence_count vs transition_count)
✅ Image pull policy corrected for :latest tags

## Related Fixes (Completed by Agents)
- RabbitMQ vhost lifecycle hook added (commit 1e66941)
- scraper-orchestrator image built and pushed to registry
- Documentation and verification scripts created

## Testing
- TypeScript: 0 syntax errors (verified)
- Python: Syntax validation passed
- Docker images: Built and pushed successfully

**Note**: K3s restart required to resolve containerd issue blocking new pod starts

+++ services/rest_api/routers/pathfinder.py
@@ -147,7 +147,7 @@ def calculate_heuristic(
     current_duration: int,
     target_duration: int,
     remaining_waypoints: Set[str],
-    avg_track_duration: int
+    avg_track_duration: float
 ) -> float:
     """
     Improved heuristic for A* search (2025 best practices)
@@ -402,7 +402,7 @@ def find_path(
             new_remaining_waypoints = current_state.remaining_waypoints - {neighbor_id}
 
             # Calculate heuristic
-            new_heuristic = calculate_heuristic(new_duration, target_duration, new_remaining_waypoints, avg_duration)
+            new_heuristic = calculate_heuristic(new_duration, target_duration, new_remaining_waypoints, float(avg_duration))
 
             # Create new state
             new_state = PathfinderState(
